server/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
click==8.1.3
constantly==15.1.0
coverage==7.1.0
cryptography==39.0.0
daphne==4.0.0
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hiredis==2.1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
more-itertools==8.14.0
pptree==3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-ulid==1.1.0
pytz==2022.7.1
redis==4.4.2
redis-om==0.1.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
types-pyOpenSSL==23.0.0.2
types-redis==4.4.0.3
typing_extensions==4.4.0
zope.interface==5.5.2
